Fix process.env logic
