+++ src/server/controller/game.ts
@@ -1,4 +1,3 @@
-import path = require("path");
 import express = require("express");
 import { resolveEngineFilesVariable } from "../utils";
 
@@ -9,14 +8,7 @@ const controller: express.RequestHandler = (req: express.Request, res: express.R
 	// json の読み込みのため require の lint エラーを抑止
 	// eslint-disable-next-line @typescript-eslint/no-var-requires
 	const pkgJson = require("../../package.json");
-
-	let engineFilesVariable: string = "";
-	if (process.env.ENGINE_FILES_V3_PATH) {
-		const filename = path.basename(process.env.ENGINE_FILES_V3_PATH, ".js");
-		engineFilesVariable = filename.replace(/[\.-]/g, "_");
-	} else {
-		engineFilesVariable = resolveEngineFilesVariable(version);
-	}
+	const engineFilesVariable = resolveEngineFilesVariable(version);
 
 	res.render("game", {
 		title: `akashic-sandbox v${pkgJson.version}`,
